+++ Client/Pages/Room.razor.cs
@@ -734,7 +734,6 @@ namespace FreeTabletop.Client.Pages
                         Tabletop.Cells[cells[i].index].Style = cells[i].style;
                         Hub.ChangeCellStyle(cells[i].index, cells[i].style);
                     }
-                    
                 }
             }
         }
@@ -742,6 +741,29 @@ namespace FreeTabletop.Client.Pages
         public void CenterOnPawn()
         {
             JSRuntime.InvokeVoidAsync("LocatePawn");
+        }
+
+        public void ToggleBleeding(Entity entity)
+        {
+            if (entity.IsBleeding)
+            {
+                entity.IsBleeding = false;
+            }
+            else
+            {
+                entity.IsBleeding = true;
+            }
+            Hub.SetBleeding(entity.UID, entity.IsBleeding);
+        }
+
+        public void SetEntityBleeding(string uid, bool isBleeding)
+        {
+            Entity entity = Tabletop.GetEntityByUID(uid);
+            if (entity != null)
+            {
+                entity.IsBleeding = isBleeding;
+                StateHasChanged();
+            }
         }
     }
 }