+++ Client/wwwroot/js/tabletop.js
@@ -94,7 +94,9 @@ class Tabletop extends HTMLElement {
                         this.cells[i].style = "fog";
                         break;
                     case "Highlighter":
-                        this.cells[i].style = "highlight";
+                        if (this.cells[i].style !== "fog" || this.isGM) {
+                            this.cells[i].style = "highlight";
+                        }
                         break;
                     default:
                         if (this.isGM || this.cells[i].style === "highlight") {
@@ -121,7 +123,7 @@ class Tabletop extends HTMLElement {
                             this.ctx.fillStyle = "rgba(255, 13, 65, 0.15)";
                             break;
                         case "fog":
-                            this.ctx.fillStyle = `rgba(25,25,25,${this.isGM ? "0.6" : "1"})`;
+                            this.ctx.fillStyle = `rgba(100,100,100,${this.isGM ? "0.6" : "1"})`;
                             break;
                         default:
                             this.ctx.fillStyle = "transparent";