+++ Shared/Models/Tabletop.cs
@@ -22,5 +22,47 @@ namespace FreeTabletop.Shared.Models
         public int[] Size { get; set; }
         public List<Cell> Cells = new List<Cell>();
         public string MessageUID {get; set;}
+
+        public Entity GetEntityByUID(string uid)
+        {
+            Entity entity = null;
+            bool FoundEntity = false;
+            for (int i = 0; i < NPCs.Count; i++)
+            {
+                if (NPCs[i].UID == uid)
+                {
+                    FoundEntity = true;
+                    entity = NPCs[i];
+                    break;
+                }
+            }
+            for (int i = 0; i < Creatures.Count; i++)
+            {
+                if (FoundEntity)
+                {
+                    break;
+                }
+                if (Creatures[i].UID == uid)
+                {
+                    FoundEntity = true;
+                    entity = Creatures[i];
+                    break;
+                }
+            }
+            for (int i = 0; i < Players.Count; i++)
+            {
+                if (FoundEntity)
+                {
+                    break;
+                }
+                if (Players[i].UID == uid)
+                {
+                    FoundEntity = true;
+                    entity = Players[i];
+                    break;
+                }
+            }
+            return entity;
+        }
     }
 }