+++ changelog.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- bleeding VFX
+
+### Fixed
+
+- adjusted cells fog color
+
 ## [0.5.1] - 2020-12-22
 
 ### Fixed