+++ changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.6.0] - 2020-12-29
+
 ### Added
 
 - bleeding effect
@@ -147,6 +149,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - initial beta build
 - CI/CD pipeline
 
+[0.6.0]: https://github.com/codewithkyle/free-tabletop-app/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/codewithkyle/free-tabletop-app/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/codewithkyle/free-tabletop-app/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/codewithkyle/free-tabletop-app/compare/v0.3.4...v0.4.0